+++ src/Features/Navigation/Search/useManualSearchParamChange.js
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+
+export const useManualSearchParamChange = (
+  location,
+  dispatch,
+  query,
+  onSearchChange,
+  searchParams
+) => {
+  useEffect(() => {
+    const newQuery = searchParams.get("search") || "";
+    if (newQuery !== query) {
+      dispatch(onSearchChange(newQuery));
+    }
+  }, [location.search]);
+};
