+++ src/Features/Navigation/Search/useReplaceSearchParam.js
@@ -1,0 +1,23 @@
+import { useEffect } from "react";
+
+export const useReplaceSearchParam = (
+  location,
+  query,
+  history,
+  dispatch,
+  fetchMovies,
+  searchParams
+) => {
+  useEffect(() => {
+    if (query === "") {
+      searchParams.delete("search");
+      searchParams.set("page", "1");
+    } else {
+      searchParams.set("search", query);
+      searchParams.set("page", "1");
+    }
+
+    history.replace(`${location.pathname}?${searchParams.toString()}`);
+    dispatch(fetchMovies());
+  }, [query]);
+};
